requirements.txt
numpy
pandas
opencv-python-headless
pytesseract
Pillow
tqdm
lxml
scikit-learn
transformers>=4.30.0
datasets
accelerate
evaluate
torch
seqeval
requests
